Update generate_html.py
Modify to grab data from a youtube rss feed. 
I know for a fact that this AI generated iteration is broken in some ways, but I'm about to head off to class and so don't have time to fix it.

Besides, I'm curious to see how badly the robot's version of this breaks right out of the box.